Selectively rewrite sender addresses
The mail relay server used to pass through the return path for other
than mailing list mails. This meant that there should be an SPF record
corresponding to all of the return path addresses.

Instead, it was decided to rewrite the sender address so that the relay
server appears to be the sender. Rewriting is only done for
'envelope-from' (corresponding 'F' in exim4 configuration) and 'sender'
(s) fields.

The address rewriting is not done for the domains in exim4_dkim_domains,
but for all other addresses the sender address is rewritten to 'bounces'
on the primary hostname of the relay server.

This patch reuses exim4_dkim_domains since it was the same set of
domains that needed special handling for rewriting.

This patch was not tested with mailman yet since mailman was running on
another server during the setup. It may require further configuration.